+++ github3/repos/__init__.py
@@ -4,6 +4,8 @@ github3.repos
 
 This module contains the classes relating to repositories.
 
+See also: http://developer.github.com/v3/repos/
+
 """
 
 from .repo import Repository